Fixed some compiler warnings on Solaris 11
* Passing char to int argument to isxdigit(3)
* Missing newline in end-of-file string_lib.h

Ticket: None
Changelog: None

+++ lib/leech.c
@@ -127,7 +127,7 @@ static bool Purge(const LCH_Instance *const instance) {
 
     bool is_block_id = true;
     for (const char *ch = filename; *ch != '\0'; ch++) {
-      if (isxdigit(*ch) == 0) {
+      if (isxdigit((int)(*ch)) == 0) {
         is_block_id = false;
         break;
       }
